Rename of cuda to pycuda.
