+++ test/test_cuda_functions.py
@@ -2,7 +2,7 @@ import numpy as np
 from .context import skip_if_no_cuda
 
 import pytest
-from kernel_tuner import cuda
+from kernel_tuner import pycuda
 from kernel_tuner.core import KernelSource, KernelInstance
 
 try:
@@ -21,7 +21,7 @@ def test_ready_argument_list():
 
     arguments = [c, a, b]
 
-    with cuda.CudaFunctions(0) as dev:
+    with pycuda.CudaFunctions(0) as dev:
         gpu_args = dev.ready_argument_list(arguments)
 
         assert isinstance(gpu_args[0], pycuda.driver.DeviceAllocation)
@@ -44,7 +44,7 @@ def test_compile():
     kernel_name = "vector_add"
     kernel_sources = KernelSource(kernel_name, kernel_string, "cuda")
     kernel_instance = KernelInstance(kernel_name, kernel_sources, kernel_string, [], None, None, dict(), [])
-    with cuda.CudaFunctions(0) as dev:
+    with pycuda.CudaFunctions(0) as dev:
         try:
             dev.compile(kernel_instance)
         except Exception as e:
